Extract Drawer Components and Refactor
