+++ src/pages/Journal/Navigator/Footer/PageFooter.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { surface2 } from "../../../common/styles";
+import { surface2 } from "../../../../common/styles";
 
 export const PageFooter = styled.div`
   grid-area: page-footer;
