+++ src/pages/Journal/Navigator/Footer/SectionFooter.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { surface2 } from "../../../common/styles";
+import { surface2 } from "../../../../common/styles";
 
 export const SectionFooter = styled.div`
   grid-area: section-footer;
